requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==3.6.4.0
binaryornot==0.4.4
celery==5.2.7
celery-singleton==0.3.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
config==0.5.1
cookiecutter==2.6.0
crispy-bootstrap4==2024.1
crispy-tailwind==1.0.3
cron-descriptor==1.4.3
cryptography==42.0.7
csscompressor==0.9.5
datadispatch==1.0.0
defusedxml==0.8.0rc2
Django==5.0.6
django-appconf==1.0.6
django-browser-reload==1.12.1
django-cachalot==2.6.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cleanup==8.1.0
django-compressor==4.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.4.2
django-dramatiq==0.11.6
django-filter==24.2
django-hashedfilenamestorage==2.4.1
django-multiupload==0.6.1
django-redis==5.2.0
django-tailwind==3.8.0
django-templated-mail==1.1.1
django-timezone-field==6.1.0
dramatiq==1.17.0
dramatiq-dashboard==0.4.0
flower==2.0.1
fontawesomefree==6.5.1
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
humanize==4.9.0
idna==3.7
Jinja2==3.1.3
kombu==5.3.7
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pika==1.3.2
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pytailwindcss==0.2.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.4
redis-cli==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rjsmin==1.2.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sorl-thumbnail==12.10.0
sqlparse==0.5.0
tailwind==3.1.5b0
tailwindcss==0.0.1
tasks==2.8.0
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
waitress==3.0.0
watchdog==4.0.1
watchdog_gevent==0.2.0
wcwidth==0.2.13
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
